requirements.txt
anyio==4.9.0
attrs==21.2.0
Automat==20.2.0
Babel==2.8.0
bcrypt==3.2.0
blinker==1.9.0
boto3==1.20.34
botocore==1.23.34
brz-etckeeper==0.0.0
certifi==2020.6.20
chardet==4.0.0
click==8.1.8
colorama==0.4.4
command-not-found==0.3
configobj==5.0.6
constantly==15.1.0
cryptography==3.4.8
dbus-python==1.2.18
distlib==0.3.4
distro==1.7.0
distro-info==1.1+ubuntu0.2
docker==5.0.3
docker-compose==1.29.2
dockerpty==0.4.1
docopt==0.6.2
exceptiongroup==1.2.2
fail2ban==0.11.2
filelock==3.6.0
Flask==3.1.0
flask-cors==5.0.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.8
httplib2==0.20.2
httpx==0.25.2
hyperlink==21.0.0
idna==3.3
importlib-metadata==4.6.4
incremental==21.3.0
iotop==0.6
itsdangerous==2.2.0
jeepney==0.7.1
Jinja2==3.1.6
jmespath==0.10.0
jsonpatch==1.32
jsonpointer==2.0
jsonschema==3.2.0
keyring==23.5.0
launchpadlib==1.10.16
lazr.restfulclient==0.14.4
lazr.uri==1.0.6
MarkupSafe==3.0.2
mercurial==6.1.1
more-itertools==8.10.0
nagiosplugin==1.3.2
netifaces==0.11.0
oauthlib==3.2.0
packaging==21.3
pexpect==4.8.0
platformdirs==2.5.1
ptyprocess==0.7.0
pyasn1==0.4.8
pyasn1-modules==0.2.1
PyGObject==3.42.1
PyHamcrest==2.0.2
pyinotify==0.9.6
PyJWT==2.3.0
PyMySQL==1.0.2
pyOpenSSL==21.0.0
pyparsing==2.4.7
pyrsistent==0.18.1
pyserial==3.5
python-apt==2.4.0+ubuntu4
python-dateutil==2.8.1
python-debian==0.1.43+ubuntu1.1
python-dotenv==0.19.2
python-magic==0.4.24
python-telegram-bot==20.7
pytz==2022.1
PyYAML==5.4.1
requests==2.25.1
s3transfer==0.5.0
SecretStorage==3.3.1
service-identity==18.1.0
six==1.16.0
sniffio==1.3.1
sos==4.7.2
ssh-import-id==5.11
supervisor==4.2.5
systemd-python==234
texttable==1.6.4
Twisted==22.1.0
typing_extensions==4.13.2
ubuntu-drivers-common==0.0.0
ubuntu-pro-client==8001
ufw==0.36.1
unattended-upgrades==0.1
urllib3==1.26.5
virtualenv==20.13.0+ds
wadllib==1.3.6
websocket-client==1.2.3
Werkzeug==3.1.3
xkit==0.0.0
zipp==1.0.0
zope.interface==5.4.0
